Remove some integration tests (#4716)
* Refactor integration tests and remove netstandard tests

* csproj

* fix

* fix
