+++ src/client/Microsoft.Identity.Client/Properties/InternalsVisibleTo.cs
@@ -12,7 +12,6 @@ using Microsoft.Identity.Client;
 [assembly: InternalsVisibleTo("Microsoft.Identity.Test.SideBySide" + KeyTokens.MSAL)]
 [assembly: InternalsVisibleTo("Microsoft.Identity.Test.Integration.NetCore" + KeyTokens.MSAL)]
 [assembly: InternalsVisibleTo("Microsoft.Identity.Test.Integration.NetFx" + KeyTokens.MSAL)]
-[assembly: InternalsVisibleTo("Microsoft.Identity.Test.Integration.NetStandard" + KeyTokens.MSAL)]
 [assembly: InternalsVisibleTo("Microsoft.Identity.Test.Performance" + KeyTokens.MSAL)]
 
 [assembly: InternalsVisibleTo("CommonCache.Test.Common" + KeyTokens.MSAL)]
