+++ tests/Microsoft.Identity.Test.Integration.netcore/HeadlessTests/CancellationTests.cs
@@ -1,22 +1,15 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved. 
 // Licensed under the MIT License.
 
-#if NET_CORE // run these only on .net core as they do not hit any platform specific part
-
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
 using System.Net.Sockets;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client;
-using Microsoft.Identity.Client.Internal;
 using Microsoft.Identity.Client.Internal.Logger;
 using Microsoft.Identity.Client.Platforms.netcore;
 using Microsoft.Identity.Test.Common.Core.Helpers;
-using Microsoft.Identity.Test.Integration.Infrastructure;
 using Microsoft.Identity.Test.Integration.Infrastructure;
 using Microsoft.Identity.Test.Unit;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -117,5 +110,4 @@ namespace Microsoft.Identity.Test.Integration.HeadlessTests
         }
     }
 }
-#endif
