+++ tests/Microsoft.Identity.Test.Integration.netcore/HeadlessTests/WamLoggerValidator.cs
