+++ tests/Microsoft.Identity.Test.Integration.netcore/Infrastructure/ClientAppTestExtensions.cs
