+++ tests/Microsoft.Identity.Test.Integration.netcore/Infrastructure/CoreUiTestConstants.cs
