+++ tests/Microsoft.Identity.Test.Integration.netcore/Infrastructure/MsalAssert.cs
