+++ tests/Microsoft.Identity.Test.Integration.netcore/Infrastructure/MsalExtensions.cs
