+++ tests/Microsoft.Identity.Test.Integration.netcore/Infrastructure/PoPValidator.cs
@@ -28,16 +28,16 @@ namespace Microsoft.Identity.Test.Integration.Infrastructure
         /// This calls a special endpoint that validates any POP token against a configurable HTTP request.
         /// The HTTP request is configured through headers.
         /// </summary>
-        public static Task VerifyPoPTokenAsync(
+        public static void VerifyPoPToken(
             string clientId, 
             string requestUri, 
             HttpMethod method, 
             AuthenticationResult result)
         {
-            return VerifyPoPTokenAsync(clientId, requestUri, method, result.AccessToken, result.TokenType);
+            VerifyPoPToken(clientId, requestUri, method, result.AccessToken, result.TokenType);
         }
 
-        public static async Task VerifyPoPTokenAsync(
+        public static void VerifyPoPToken(
             string clientId, 
             string requestUri, 
             HttpMethod method, 
@@ -60,7 +60,7 @@ namespace Microsoft.Identity.Test.Integration.Infrastructure
             Assert.IsNotNull(cnf);
             ClaimsPrincipal innerTokenClaims = IdToken.Parse(assertionWithoutShr).ClaimsPrincipal;
             string reqCnf = innerTokenClaims.FindFirst("cnf").Value;
-            Assert.IsNotNull(reqCnf);            
+            Assert.IsNotNull(reqCnf);
         }
 
         public static async Task VerifyPopNonceAsync(string nonce)
