+++ tests/Microsoft.Identity.Test.Integration.netcore/Infrastructure/SeleniumExtensions.cs
