+++ tests/Microsoft.Identity.Test.Integration.netcore/Infrastructure/SeleniumWebUI.cs
