+++ tests/Microsoft.Identity.Test.Integration.netcore/Infrastructure/UserInformationFieldIds.cs
