+++ tests/Microsoft.Identity.Test.Integration.netcore/Microsoft.Identity.Test.Integration.NetCore.csproj
@@ -9,8 +9,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="../Microsoft.Identity.Test.Integration.NetFx/**/*.cs" Exclude="../Microsoft.Identity.Test.Integration.NetFx/obj/**/*" />
-    <None Include="../Microsoft.Identity.Test.Integration.NetFx/federationMetadata.xml" />
     <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client.Broker\Microsoft.Identity.Client.Broker.csproj" />
     <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client.Extensions.Msal\Microsoft.Identity.Client.Extensions.Msal.csproj" />
     <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client\Microsoft.Identity.Client.csproj" />
@@ -37,10 +35,11 @@
     <!-- Workaround for hitting exception invoking executor 'executor://nunit3testexecutor/': Could not load type 'System.Runtime.Remoting.Channels.IChannel'-->
     <PackageReference Include="NUnit" />
     <PackageReference Include="NUnit3TestAdapter" />
+    <PackageReference Include="StrongNamer" />
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="..\Microsoft.Identity.Test.Integration.NetFx\federationMetadata.xml">
+    <None Update="federationMetadata.xml">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
