+++ tests/Microsoft.Identity.Test.Integration.netcore/SeleniumTests/ConfidentialClientAuthorizationTests.cs
@@ -34,7 +34,6 @@ namespace Microsoft.Identity.Test.Integration.SeleniumTests
 
         private static readonly string[] s_scopes = { "User.Read" };
         private const string ConfidentialClientID = "35dc5034-9b65-4a5d-ad81-73cca468c1e0"; //msidlab4.com app
-        private const string TenantId = "72f988bf-86f1-41af-91ab-2d7cd011db47";
         private const string CertificateName = "for-cca-testing";
 
         private static KeyVaultSecretsProvider s_secretProvider;
@@ -59,8 +58,7 @@ namespace Microsoft.Identity.Test.Integration.SeleniumTests
 
         #endregion
 
-        [TestMethod]
-        // Regression test for: https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/891
+        [TestMethod]        // Regression test for: https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/891
         public async Task SeleniumGetAuthCode_RedeemForAt_CommonAuthority_Async()
         {
             // Arrange
