+++ tests/Microsoft.Identity.Test.Integration.netcore/federationMetadata.xml
