+++ tests/Microsoft.Identity.Test.Integration.netfx/Microsoft.Identity.Test.Integration.NetFx.csproj
@@ -6,7 +6,14 @@
     <Configurations>Debug;Release;Debug + MobileApps</Configurations>
   </PropertyGroup>
 
+  <ItemGroup Label="Only a few tests run on .NET FWK. Define the files here">
+    <Compile Include="../Microsoft.Identity.Test.Integration.NetCore/**/*.NetFwk.cs" Exclude="../Microsoft.Identity.Test.Integration.NetFx/obj/**/*" />
+    <Compile Include="../Microsoft.Identity.Test.Integration.NetCore/Infrastructure/**/*.cs" LinkBase="Infrastructure"></Compile>
+  
+  </ItemGroup>
+  
   <ItemGroup>
+    
     <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client.Extensions.Msal\Microsoft.Identity.Client.Extensions.Msal.csproj" />
     <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client\Microsoft.Identity.Client.csproj" />
     <ProjectReference Include="..\Microsoft.Identity.Test.LabInfrastructure\Microsoft.Identity.Test.LabInfrastructure.csproj" />
@@ -31,13 +38,9 @@
 
     <PackageReference Include="coverlet.collector" />
   </ItemGroup>
-
-  <ItemGroup>
-    <None Update="federationMetadata.xml">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
   
+  <ItemGroup>
+    <Folder Include="Infrastructure\" />
+  </ItemGroup>
   
 </Project>
