--- tests/Microsoft.Identity.Test.Integration.netstandard/Microsoft.Identity.Test.Integration.NetStandard.csproj
@@ -1,59 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <IsPackable>false</IsPackable>
-    <DefineConstants>$(DefineConstants);NETSTANDARD</DefineConstants>
-    <Configurations>Debug;Release;Debug + MobileApps</Configurations>
-  </PropertyGroup>
-
-  <ItemGroup>
-
-    <Compile Include="../Microsoft.Identity.Test.Integration.NetFx/**/*.cs" Exclude="../Microsoft.Identity.Test.Integration.NetFx/obj/**/*" />
-
-    <Compile Remove="..\Microsoft.Identity.Test.Integration.NetFx\HeadlessTests\OtelTests.cs" />
-    <None Include="../Microsoft.Identity.Test.Integration.NetFx/federationMetadata.xml" />
-
-    <!-- Force pick the netstandard version of our assemblies -->
-    <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client.Broker\Microsoft.Identity.Client.Broker.csproj" AdditionalProperties="TargetFramework=netstandard2.0" />
-    <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client.Extensions.Msal\Microsoft.Identity.Client.Extensions.Msal.csproj" AdditionalProperties="TargetFramework=netstandard2.0" />    
-    <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client\Microsoft.Identity.Client.csproj" AdditionalProperties="TargetFramework=netstandard2.0" />
-    <ProjectReference Include="..\Microsoft.Identity.Test.LabInfrastructure\Microsoft.Identity.Test.LabInfrastructure.csproj" />
-    <ProjectReference Include="..\Microsoft.Identity.Test.Common\Microsoft.Identity.Test.Common.csproj" />
-    
-    <PackageReference Include="Microsoft.AspNet.WebApi.Client" />
-    <PackageReference Include="Microsoft.IdentityModel.JsonWebTokens" />
-    <PackageReference Include="Microsoft.IdentityModel.Protocols.SignedHttpRequest" />
-    <PackageReference Include="Newtonsoft.Json" />
-    <PackageReference Include="Selenium.Support" />
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" />
-
-    <!-- This reference is a workaround for a bug in .net46
-    https://stackoverflow.com/questions/45563560/could-not-load-file-or-assembly-system-net-http-version-4-1-1-1-net-standard-->
-    <PackageReference Include="System.Net.Http" />
-    <PackageReference Include="Selenium.WebDriver" />
-    <PackageReference Include="StrongNamer" />
-    <PackageReference Include="System.Text.Json" />
-
-    <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="MSTest.TestAdapter" />
-    <PackageReference Include="MSTest.TestFramework" />
-    <PackageReference Include="coverlet.collector" />
-
-    <PackageReference Include="OpenTelemetry.Exporter.Console" />
-    <PackageReference Include="OpenTelemetry.Exporter.InMemory" />
-
-    <!-- Workaround for hitting exception invoking executor 'executor://nunit3testexecutor/': Could not load type 'System.Runtime.Remoting.Channels.IChannel'-->
-    <PackageReference Include="NUnit" />
-    <PackageReference Include="NUnit3TestAdapter" />
-    
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="..\Microsoft.Identity.Test.Integration.NetFx\federationMetadata.xml">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-  
-</Project>
